refactor: rename presentation module to tui
