+++ cli/src/templates/AGENTS.md
@@ -4,10 +4,10 @@ This project uses **ucharm** - a CLI toolkit for building beautiful command-line
 
 ## Critical Context
 
-- **Runtime**: PocketPy with native Zig modules (NOT CPython)
+- **Runtime**: PocketPy with native Rust modules (NOT CPython)
 - **No pip packages**: Cannot use packages with C extensions (no requests, numpy, pandas)
 - **Output**: Standalone binaries (~1MB)
-- **50+ runtime modules** including: ansi, args, argparse, base64, charm, collections, copy, csv, dataclasses, datetime, fetch, fnmatch, functools, glob, gzip, hashlib, heapq, hmac, http.client, input, itertools, json, logging, math, operator, os, pathlib, random, re, secrets, shutil, signal, sqlite3, statistics, struct, subprocess, tarfile, tempfile, template, term, textwrap, time, toml, typing, unittest, urllib.parse, uuid, xml.etree.ElementTree, zipfile
+- **50+ runtime modules** including: ansi, args, argparse, base64, tui, collections, copy, csv, dataclasses, datetime, fetch, fnmatch, functools, glob, gzip, hashlib, heapq, hmac, http.client, input, itertools, json, logging, math, operator, os, pathlib, random, re, secrets, shutil, signal, sqlite3, statistics, struct, subprocess, tarfile, tempfile, template, term, textwrap, time, toml, typing, unittest, urllib.parse, uuid, xml.etree.ElementTree, zipfile
 
 ## Import Pattern
 
@@ -16,7 +16,7 @@ This project uses **ucharm** - a CLI toolkit for building beautiful command-line
 from ucharm import box, success, error, select, confirm, prompt
 
 # Or import native modules directly
-import charm
+import tui
 import input
 ```
