+++ cli/src/templates/CLAUDE.md
@@ -4,15 +4,15 @@ This project uses **ucharm** - a CLI toolkit for building beautiful command-line
 
 ## Key Concepts
 
-- **PocketPy, not CPython**: This runs on PocketPy with native Zig modules, not standard Python
-- **Native modules**: 50+ high-performance modules implemented in Zig (see list below)
+- **PocketPy, not CPython**: This runs on PocketPy with native Rust modules, not standard Python
+- **Native modules**: 50+ high-performance modules implemented in Rust (see list below)
 - **No pip packages**: You cannot use pip packages that have C extensions
 - **Single binary output**: Apps compile to standalone executables (~1MB)
 
 ## Available Modules
 
 ### TUI Components
-- `charm` - Box, table, rule, progress bar, spinner, status messages (success/error/warning/info)
+- `tui` - Box, table, rule, progress bar, spinner, status messages (success/error/warning/info)
 - `input` - Interactive prompts: select, multiselect, confirm, prompt, password
 - `term` - Terminal control (size, raw mode, cursor, colors)
 - `ansi` - ANSI escape codes for styling
@@ -40,7 +40,7 @@ This project uses **ucharm** - a CLI toolkit for building beautiful command-line
 from ucharm import box, success, select, confirm
 
 # Or import native modules directly
-import charm
+import tui
 import input
 ```
