+++ tests/compat_report_pocketpy.md
@@ -1,6 +1,6 @@
 # μcharm Compatibility Report
 
-Generated: 2026-08-04 23:36:22
+Generated: 2026-08-05 00:24:21
 
 ## Summary
 
@@ -277,5 +277,5 @@ The following 109 CPython stdlib modules are not yet targeted:
 
 - Tests are adapted from CPython's test suite
 - Some tests require features not available in PocketPy (threading, gc introspection)
-- μcharm-specific modules (ansi, charm, input, term, args) have custom tests
+- μcharm-specific modules (ansi, tui, input, term, args) have custom tests
 - Report generated by `python3 tests/compat_runner.py --report`
