+++ tests/compat_runner.py
@@ -352,7 +352,7 @@ CPYTHON_STDLIB_ALL = {
 UCHARM_MODULES = [
     "ansi",  # ANSI color codes
     "args",  # CLI argument parsing
-    "charm",  # TUI components (box, rule, progress, etc.)
+    "tui",  # TUI components (box, rule, progress, etc.)
     "input",  # Interactive input (select, confirm, prompt)
     "path",  # Path manipulation utilities
     "term",  # Terminal control
@@ -624,7 +624,7 @@ def run_all_tests(
         result = test_module(module, "stdlib", test_dir, mpy_path, verbose)
         results.append(result)
 
-    # Note: ucharm-specific modules (ansi, charm, input, term, args, path) are
+    # Note: ucharm-specific modules (ansi, tui, input, term, args, path) are
     # our own libraries - no CPython equivalent exists, so no comparison needed.
     # They are accepted as-is.
 
@@ -1007,7 +1007,7 @@ def generate_report(results: list[ModuleResult], output_path: Path):
             "",
             "- Tests are adapted from CPython's test suite",
             "- Some tests require features not available in PocketPy (threading, gc introspection)",
-            "- μcharm-specific modules (ansi, charm, input, term, args) have custom tests",
+            "- μcharm-specific modules (ansi, tui, input, term, args) have custom tests",
             "- Report generated by `python3 tests/compat_runner.py --report`",
         ]
     )
