+++ website/content/docs/modules/ansi.mdx
@@ -12,7 +12,7 @@ import ansi
 ```
 
 <Callout>
-For most use cases, prefer the higher-level `charm.style()` function. Use `ansi` when you need fine-grained control or are building custom components.
+For most use cases, prefer the higher-level `tui.style()` function. Use `ansi` when you need fine-grained control or are building custom components.
 </Callout>
 
 ## Colors
