+++ website/content/docs/modules/index.mdx
@@ -5,13 +5,13 @@ description: Overview of all ucharm runtime modules
 
 # Modules
 
-ucharm includes 50+ native modules implemented in Zig for maximum performance and minimal binary size.
+ucharm includes 50+ native modules implemented in Rust for performance, safety, and a compact binary.
 
 ## TUI Components
 
 | Module | Description |
 |--------|-------------|
-| [charm](/docs/modules/charm) | Boxes, tables, progress bars, spinners, status messages |
+| [tui](/docs/modules/tui) | Boxes, tables, progress bars, spinners, status messages |
 | [input](/docs/modules/input) | Interactive prompts: select, multiselect, confirm, password |
 | [term](/docs/modules/term) | Terminal control: size, cursor, raw mode |
 | [ansi](/docs/modules/ansi) | ANSI escape codes for colors and styles |
